Extract dependency check into a custom plugin
