+++ buildSrc/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins {
+    `kotlin-dsl`
+    `java-gradle-plugin`
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    api(gradleApi())
+}
+
+tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
+    kotlinOptions {
+        jvmTarget = "11"
+    }
+}
