+++ buildSrc/src/main/resources/META-INF/gradle-plugins/org.eclipse.dataspaceconnector.gradle.properties
@@ -1,0 +1,1 @@
+implementation-class=org.eclipse.dataspaceconnector.gradle.EDCGradlePlugin
